refactor(core): create core.permanences subpackage
