+++ tipi/core/permanences/__init__.py
@@ -1,0 +1,28 @@
+"""
+Core permanences and loggers for TensorImagePipeline.
+
+This package contains all permanence classes (Device, ProgressManager) and logger implementations.
+"""
+
+from tipi.core.permanences.device import Device, DeviceWithVRAM, VRAMUsageError
+from tipi.core.permanences.loggers import (
+    BaseLoggerManager,
+    BasicLogger,
+    NullWandBLogger,
+    TensorBoardLogger,
+    WandBLogger,
+)
+from tipi.core.permanences.progress import NullProgressManager, ProgressManager
+
+__all__ = [
+    "BaseLoggerManager",
+    "BasicLogger",
+    "Device",
+    "DeviceWithVRAM",
+    "NullProgressManager",
+    "NullWandBLogger",
+    "ProgressManager",
+    "TensorBoardLogger",
+    "VRAMUsageError",
+    "WandBLogger",
+]
