+++ tipi/core/permanences/loggers/__init__.py
@@ -1,0 +1,12 @@
+from tipi.core.permanences.loggers.base import BaseLoggerManager
+from tipi.core.permanences.loggers.basic import BasicLogger
+from tipi.core.permanences.loggers.tensorboard import TensorBoardLogger
+from tipi.core.permanences.loggers.wandb import NullWandBLogger, WandBLogger
+
+__all__ = [
+    "BaseLoggerManager",
+    "BasicLogger",
+    "NullWandBLogger",
+    "TensorBoardLogger",
+    "WandBLogger",
+]
