+++ tipi/core/permanences/loggers/base.py
