+++ tipi/core/permanences/loggers/tensorboard.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Any
 
-from tipi.core.loggers.base import BaseLoggerManager
+from tipi.core.permanences.loggers.base import BaseLoggerManager
 
 
 class TensorBoardLogger(BaseLoggerManager):
